Moved file handling out of directory
