+++ FieldOpt/Reservoir/tests/well_index_calculation/test_intersected_cells.cpp
@@ -4,7 +4,7 @@
 #include "Reservoir/grid/eclgrid.h"
 #include "Reservoir/grid/grid_exceptions.h"
 #include "Reservoir/well_index_calculation/wellindexcalculator.h"
-#include "Utilities/file_handling/filehandling.hpp"
+#include "Utilities/filehandling.hpp"
 
 using namespace Reservoir::Grid;
 using namespace Reservoir::WellIndexCalculation;
