+++ FieldOpt/Simulation/simulator_interfaces/driver_file_writers/flowdriverfilewriter.cpp
@@ -1,7 +1,7 @@
 #include <simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/welspecs.h>
 #include <simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/compdat.h>
 #include <simulator_interfaces/driver_file_writers/driver_parts/ecl_driver_parts/wellcontrols.h>
-#include <Utilities/file_handling/filehandling.hpp>
+#include <Utilities/filehandling.hpp>
 #include "flowdriverfilewriter.h"
 
 namespace Simulation {
