+++ FieldOpt/Utilities/settings/model.cpp
@@ -26,7 +26,7 @@
 #include <iostream>
 #include "Utilities/settings/model.h"
 #include "settings_exceptions.h"
-#include "Utilities/file_handling/filehandling.hpp"
+#include "filehandling.hpp"
 
 namespace Utilities {
     namespace Settings {
