+++ FieldOpt/Utilities/settings/settings.cpp
@@ -1,6 +1,6 @@
 #include "settings.h"
 #include "settings_exceptions.h"
-#include "Utilities/file_handling/filehandling.hpp"
+#include "filehandling.hpp"
 
 #include <QJsonDocument>
 #include <iostream>
