+++ FieldOpt/Utilities/tests/file_handling/test_filehandling.cpp
@@ -25,7 +25,7 @@
 
 #include <gtest/gtest.h>
 
-#include "Utilities/file_handling/filehandling.hpp"
+#include "filehandling.hpp"
 #include "Utilities/tests/test_resource_example_file_paths.hpp"
 
 namespace {
